requirements.txt
# Editable install with no version control (6476-Winter2021-Proj1==1.0.0)
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
cycler==0.10.0
# dataclasses==0.8
dataclasses
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
importlib-metadata==3.7.3
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
